requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
exceptiongroup==1.2.0
Flask==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.26.0
sniffio==1.3.0
SQLAlchemy==2.0.25
telegram==0.0.1
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.4
